+++ arrays/src/main/java/mathtools/arrays/IntArrayExt.java
@@ -1,4 +1,4 @@
-package mathtools.lists.arrays;
+package mathtools.arrays;
 
 import java.util.ArrayList;
 import java.util.Collections;
