+++ arrays/src/main/java/mathtools/arrays/package-info.java
@@ -2,4 +2,4 @@
  * Operations and Data Structures on Array types
  * @author DK96-OS
  */
-package mathtools.lists.arrays;
+package mathtools.arrays;
