+++ lists-ktx/src/main/kotlin/mathtools/lists/ByteList.kt
@@ -138,7 +138,7 @@ object ByteList {
 			1 -> BigInteger.valueOf(list[0].toLong())
 			else -> BigInteger.ZERO
 		}
-		val buffer = BigIntSumBuffer()
+		val buffer = mathtools.arrays.sum.BigIntSumBuffer()
 		for (i in list) buffer.add(i.toLong())
 		return buffer.sum
 	}
